+++ src/ft_utils.c
@@ -6,16 +6,16 @@
 /*   By: hokutosuzuki <hosuzuki@student.42toky      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/07 07:14:49 by hokutosuz         #+#    #+#             */
-/*   Updated: 2022/02/10 07:10:10 by hokutosuz        ###   ########.fr       */
+/*   Updated: 2022/02/10 11:33:19 by hokutosuz        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "../libft/libft.h"
 
-int	ft_intmax(t_stock *lst, long total, int len)
+int	ft_intmax(t_stock *lst, unsigned long long total, int len)
 {
-	if ((long)INT_MAX <= total + len)
+	if ((unsigned long long)INT_MAX <= total + (unsigned long long)len)
 	{
 		lst->status = ERROR;
 		return (ERROR);
